feat: add HeaderControls component and update layouts
- Introduced a new HeaderControls component for managing language selection and Fi connection status.
- Updated RootLayout to conditionally render HeaderControls based on the current route.
- Refactored AgentsLayout and other agent pages to remove HeaderControls, ensuring a cleaner layout.
- Enhanced Sidebar component to utilize user profile images and display user information more effectively.
