+++ src/app/agents/debt-squasher/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import React, { useState, useRef, useEffect } from "react";
-import { Sidebar } from "@/components/Sidebar";
+
 import Particles from "@/ui/components/Particles";
 import TextType from "@/ui/TextAnimations/TextType/TextType";
 import { useChat } from "@/hooks/useChat";
@@ -591,24 +591,10 @@ export default function DebtSquasherPage() {
   return (
     <div
       ref={containerRef}
-      className="flex flex-col lg:flex-row h-screen w-screen bg-[#111827] overflow-hidden smooth-entry"
+      className="flex flex-col h-full w-full bg-transparent overflow-hidden smooth-entry"
+      style={{ pointerEvents: "auto" }}
     >
-      {/* Header Controls */}
-      <HeaderControls />
 
-      {/* Sidebar */}
-      <div className="relative z-10 w-full lg:w-auto lg:flex-shrink-0">
-        <Sidebar
-          user={user}
-          // starredSessions={sessionsByCategory.starred}
-          // chatSessions={sessionsByCategory.chats}
-          activeSessionId={activeSessionId}
-          onSelectSession={selectSession}
-          onDeleteSession={deleteSession}
-          onNewSession={createNewSession}
-          onToggleStar={toggleSessionStar}
-        />
-      </div>
 
       {/* Particles Background - Full Screen */}
       <div className="fixed inset-0 w-screen h-screen z-0 blur-sm opacity-70">
